+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.92.0] - 2026-09-18
+
 ### Added
 
 - **PyO3: `[crates.python] send_sync_types` opts extracted opaque classes into checked thread